requirements.txt
python-docx
pandas
nltk
scikit-learn
pyttsx3
SpeechRecognition
pyaudio
genai apikey = xt